--- a/JobScheduler/JobScheduler.csproj
+++ b/JobScheduler/JobScheduler.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         
@@ -28,7 +28,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="7.0.0" />
+        <PackageReference Include="DequeNET" Version="1.0.2" />
+        <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="7.0.0" />
     </ItemGroup>